Move module Kore.Internal.Substitute to Kore.Substitute
