+++ kore/src/Kore/Equation/Simplification.hs
@@ -26,7 +26,6 @@ import Kore.Internal.MultiAnd (
 import qualified Kore.Internal.MultiAnd as MultiAnd
 import qualified Kore.Internal.Predicate as Predicate
 import qualified Kore.Internal.SideCondition as SideCondition
-import Kore.Internal.Substitute
 import qualified Kore.Internal.Substitution as Substitution
 import qualified Kore.Internal.TermLike as TermLike
 import Kore.Rewriting.RewritingVariable (
@@ -36,6 +35,7 @@ import Kore.Step.Simplification.Simplify (
     MonadSimplify,
  )
 import qualified Kore.Step.Simplification.Simplify as Simplifier
+import Kore.Substitute
 import Kore.TopBottom
 import qualified Logic
 import Prelude.Kore
