+++ kore/src/Kore/Step/Simplification/Pattern.hs
@@ -34,7 +34,6 @@ import qualified Kore.Internal.SideCondition as SideCondition (
     assumeDefined,
     top,
  )
-import Kore.Internal.Substitute
 import Kore.Internal.Substitution (
     toMap,
  )
@@ -49,6 +48,7 @@ import Kore.Step.Simplification.Simplify (
     simplifyCondition,
     simplifyConditionalTerm,
  )
+import Kore.Substitute
 import Prelude.Kore
 
 -- | Simplifies the 'Pattern' and removes the exists quantifiers at the top.
