+++ kore/src/Kore/Substitute.hs
@@ -2,7 +2,7 @@
 Copyright   : (c) Runtime Verification, 2021
 License     : NCSA
 -}
-module Kore.Internal.Substitute (
+module Kore.Substitute (
     Substitute (..),
 ) where
