+++ nix/kore.nix.d/kore.nix
@@ -242,6 +242,7 @@
           "Kore/Internal/Predicate"
           "Kore/Internal/SideCondition"
           "Kore/Internal/SideCondition/SideCondition"
+          "Kore/Substitute"
           "Kore/Internal/Substitution"
           "Kore/Internal/Symbol"
           "Kore/Internal/TermLike"
@@ -392,7 +393,6 @@
           "Kore/Step/Strategy"
           "Kore/Step/Substitution"
           "Kore/Step/Transition"
-          "Kore/Substitute"
           "Kore/Syntax"
           "Kore/Syntax/And"
           "Kore/Syntax/Application"
